requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
dnspython
python-whois
pydantic
pydantic-settings
